add docker login for ocr

+++ platform-operator/build/scripts/cleanup.sh
@@ -8,8 +8,11 @@
 echo "Doing cleanup of KIND clusters and Docker containers"
 kind delete cluster --name "$1"
 
-echo "Deleting the kubeconfig '${KUBECONFIG}' ..."
-rm ${KUBECONFIG}
+if [ -f "${KUBECONFIG}" ]
+then
+  echo "Deleting the kubeconfig '${KUBECONFIG}' ..."
+  rm ${KUBECONFIG}
+fi
 
 containers=$(docker ps -q --filter label=io.x-k8s.kind.cluster | wc -l)
 if [ "$containers" -gt 1 ]
